core: rename trezor.ui.slip39 to more general name num_input
