+++ core/src/trezor/ui/num_input.py
